[type:feat]Redesigned random data generator
