+++ shenyu-plugin/shenyu-plugin-mock/src/main/java/org/apache/shenyu/plugin/mock/generator/BoolGenerator.java
@@ -15,57 +15,35 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.plugin.mock.generator.impl;
+package org.apache.shenyu.plugin.mock.generator;
 
-import java.util.List;
-import java.util.Objects;
-import org.apache.shenyu.plugin.mock.generator.Generator;
+import org.apache.shenyu.plugin.base.mock.Generator;
+import org.apache.shenyu.plugin.mock.util.RandomUtil;
 import org.apache.shenyu.spi.Join;
 
 /**
- * Random double value generator in the specified range.
+ * Boolean Generator.
  */
 @Join
-public class RandomDoubleGenerator implements Generator<String> {
-    
-    private Double min;
-    
-    private Double max;
-    
-    private String format;
+public class BoolGenerator implements Generator<Boolean> {
     
     @Override
     public String getName() {
-        return "double";
+        return "bool";
     }
     
     @Override
-    public String generate() {
-        Double result = (Math.random() * (max - min)) + min;
-        if (format != null) {
-            return String.format(format, result);
-        }
-        return String.valueOf(result);
+    public Boolean generate() {
+        return RandomUtil.randomInt(0, 1) == 1;
     }
     
     @Override
     public int getParamSize() {
-        return 1;
-    }
-    
-    @Override
-    public void initParam(final List<String> params) {
-        String[] range = params.get(0).split("-");
-        min = Double.parseDouble(range[0]);
-        max = Double.parseDouble(range[1]);
-        if (params.size() == 2) {
-            format = Objects.equals(params.get(1), "") ? null : params.get(1);
-        }
+        return 0;
     }
     
     @Override
     public boolean match(final String rule) {
-        return rule.matches("^double\\|\\d+(?:\\.\\d+)?-\\d+(?:\\.\\d+)?.*");
+        return rule.matches("^bool$");
     }
 }
-
