+++ shenyu-plugin/shenyu-plugin-mock/src/main/resources/META-INF/shenyu/org.apache.shenyu.plugin.base.mock.Generator
@@ -13,13 +13,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-bool=org.apache.shenyu.plugin.mock.generator.impl.BoolGenerator
-current=org.apache.shenyu.plugin.mock.generator.impl.CurrentTimeGenerator
-email=org.apache.shenyu.plugin.mock.generator.impl.EmailGenerator
-en=org.apache.shenyu.plugin.mock.generator.impl.EnStringGenerator
-phone=org.apache.shenyu.plugin.mock.generator.impl.PhoneGenerator
-double=org.apache.shenyu.plugin.mock.generator.impl.RandomDoubleGenerator
-int=org.apache.shenyu.plugin.mock.generator.impl.RandomIntGenerator
-list=org.apache.shenyu.plugin.mock.generator.impl.RangeDataGenerator
-zh=org.apache.shenyu.plugin.mock.generator.impl.ZhStringGenerator
+bool=org.apache.shenyu.plugin.mock.generator.BoolGenerator
+current=org.apache.shenyu.plugin.mock.generator.CurrentTimeGenerator
+email=org.apache.shenyu.plugin.mock.generator.EmailGenerator
+en=org.apache.shenyu.plugin.mock.generator.EnStringGenerator
+phone=org.apache.shenyu.plugin.mock.generator.PhoneGenerator
+double=org.apache.shenyu.plugin.mock.generator.RandomDoubleGenerator
+int=org.apache.shenyu.plugin.mock.generator.RandomIntGenerator
+list=org.apache.shenyu.plugin.mock.generator.RangeDataGenerator
+zh=org.apache.shenyu.plugin.mock.generator.ZhStringGenerator
