+++ shenyu-plugin/shenyu-plugin-mock/src/test/java/org/apache/shenyu/plugin/mock/generator/BoolGeneratorTest.java
@@ -20,7 +20,6 @@ package org.apache.shenyu.plugin.mock.generator;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.apache.shenyu.plugin.mock.generator.impl.BoolGenerator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
